backend/content_service/pyproject.toml
[project]
name = "content_service"
version = "1.0"
description = ""
authors = [
    {name = "IvanZaycev0717",email = "ivzaycev0717@yandex.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pymongo (>=4.14.1,<5.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "grpcio (>=1.74.0,<2.0.0)",
    "grpcio-tools (>=1.74.0,<2.0.0)",
    "grpcio-reflection (>=1.74.0,<2.0.0)",
    "protobuf (>=6.32.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
